[graph] Updated reporter and tests. Ready for opening.
This commit finalizes the work on graphs by implementing a graph search
algorithm and providing an improved reporting capability that identifies
top-level vars that have no references, as well as lower-level vars that
have inbound references from top-level vars that don't get called. This
lays a lot of groundwork for more interesting graph analytics if people
feel so inclined.