[protocols] Point to defprotocol forms properly when they exist
Due to the way Clojure interacts with the JVM, the existing method of
looking up var names fails to find interfaces and protocols. This is
because a `deftype` form tha references an interface is actually
pointing to the compiled JVM bytecode, not to the Clojure form. This
commit adds in some logic that checks to see if a form in question is an
interface, and if so attempts to points to the `defprotocol` form that created
it.

This commit resolves #23.