[jvm] More work on JVM interoperability.
This commit adds some of the scaffolding necessary for our long-term JVM
interoperability needs, including handling for `deftype` and `defrecord`
generator functions.

At the moment it still fails to pick up and adequately deal with direct
construction / initialization of a new class instance, e.g. (new String)
works but (String.) doesn't resolve properly. After some fussing around
with `macroexpand` I've decided to treat the problem the same way I did
with the generator functions, which is to say that I intend on using the
name of the symbol in the form as a cue for whether or not it might be a
class constructor (i.e. does it end in a period, and if so does the
implied class name resolve to a class?)

This is starting to get close to a release point, but I want to make
sure we adequately resolve the edge cases identified in #21 before
going much further.