[bugfix] Fixing a ClassNotFoundException
Resolves #30.

This fixes a problem where a var that looked like a Java class
constructor would cause Yagni to try to resolve the class it thought
existed. Unfortunately, while `resolve` returns nil when it believes
it's dealing with a Clojure symbol, when it's targeting something it
believes to be a Java class it throws a ClassNotFoundException.

This commit wraps the resolve attempt in a try/catch block - it's ugly,
but I'm not aware of a more elegant solution at the moment.